oggstream: Use proper type for sample calculation
If we are going to return a (potentially) 64bit integer, don't use
a 32bit one for calculation, otherwise we could end up exceeding
the maximum size of a 32bit int.

+++ ext/ogg/gstoggstream.c
@@ -1570,7 +1570,7 @@ static gint64
 packet_duration_ogm (GstOggStream * pad, ogg_packet * packet)
 {
   const guint8 *data;
-  int samples;
+  gint64 samples;
   int offset;
   int n;
